fix splitted actions
