--- frontend/src/actions/download.js
@@ -1,6 +1,0 @@
-export const CHART_DOWNLOAD_STATUS_UPDATE = 'CHART_DOWNLOAD_STATUS_UPDATE';
-export const updateChartDownloadStatus = (projectId, chartDownloadStatus) => ({
-  type: CHART_DOWNLOAD_STATUS_UPDATE,
-  projectId,
-  chartDownloadStatus,
-});
