+++ frontend/src/actions/index.js
@@ -1,4 +1,3 @@
-export * from './api';
 export * from './config';
-export * from './download';
+export * from './entities';
 export * from './status';
