+++ frontend/src/actions/status.js
@@ -22,38 +22,9 @@ export const updateResultFilter = (projectId, filterKey, filterText) => ({
   filterText,
 });
 
-export const TABLE_STATE_EXPANDED_UPDATE = 'TABLE_STATE_EXPANDED_UPDATE';
-export const updateTableExpanded = (projectId, expanded) => ({
-  type: TABLE_STATE_EXPANDED_UPDATE,
+export const CHART_DOWNLOAD_STATUS_UPDATE = 'CHART_DOWNLOAD_STATUS_UPDATE';
+export const updateChartDownloadStatus = (projectId, chartDownloadStatus) => ({
+  type: CHART_DOWNLOAD_STATUS_UPDATE,
   projectId,
-  expanded,
-});
-
-export const TABLE_STATE_COLUMNS_VISIBILITY_UPDATE = 'TABLE_STATE_COLUMNS_VISIBILITY_UPDATE';
-export const updateTableColumnsVisibility = (
-  projectId,
-  knownLogKeysConfig,
-  knownArgKeysConfig,
-  isGrouped
-) => ({
-  type: TABLE_STATE_COLUMNS_VISIBILITY_UPDATE,
-  projectId,
-  knownLogKeysConfig,
-  knownArgKeysConfig,
-  isGrouped,
-});
-
-export const ASSETS_TABLE_STATE_COLUMNS_VISIBILITY_UPDATE =
-  'ASSETS_TABLE_STATE_COLUMNS_VISIBILITY_UPDATE';
-export const updateAssetsTableColumnsVisibility = (
-  projectId,
-  resultId,
-  knownTrainInfoKeysConfig,
-  knownContentKeysConfig
-) => ({
-  type: ASSETS_TABLE_STATE_COLUMNS_VISIBILITY_UPDATE,
-  projectId,
-  resultId,
-  knownTrainInfoKeysConfig,
-  knownContentKeysConfig,
+  chartDownloadStatus,
 });
